[AI][MOTION] Updated structure to really work.

+++ Eduras/src/de/illonis/eduras/ai/movement/PathFinder.java
@@ -61,6 +61,6 @@ public interface PathFinder {
 	 * 
 	 * @return total distance.
 	 */
-	double getDistance();
+	float getDistance();
 
 }
